refactor(ims): split agent.rs into a directory, decompose dispatch_loop
ims/agent.rs was 3790 lines — the worst production-code offender in the
tree — and its layout hid things: production code resumed *below* the test
module, so probe_inbound sat where nobody would look for it.

  mod.rs          entry points, serve_inbound, and the dispatch loop
  inbound.rs      answering one carrier INVITE, and the PBX ring wait
  origination.rs  the outbound call state machine
  call.rs         ActiveCall/DialogInfo and every way a call ends
  ping.rs         Gm connection liveness and repair
  veth.rs         the link to Agent B, and the RTP relay
  probe.rs        the probe-inbound diagnostic

Beyond relocation:

- dispatch_loop was a single ~590-line function with fifteen parameters
  and ten mutable locals. Its inputs are now DispatchParams, its
  carried-across-iterations state is LoopState, and the body is a loop
  over ten named methods (sync_shared_status, handle_pbx_hangup,
  handle_attachment_loss, advance_origination, handle_place_call_request,
  handle_inbound_invite, handle_carrier_bye, handle_carrier_response,
  on_idle_tick). Every `continue` in the old match arms became a `return`
  from the extracted method, which is equivalent because nothing followed
  the match in the loop body.
- send_cancel and ack_and_bye_racing_answer each took eight arguments that
  were all fields of PendingOrigination; both are now methods on it. With
  begin_origination's veth arguments bundled into OriginationSetup, all
  three #[allow(clippy::too_many_arguments)] on this path are gone.
- The BYE that ends a dialog was built by hand at four sites; it is now
  DialogInfo::build_bye_for.
- handle_invite's nine parameters become InviteContext.

Behaviour is unchanged: 1270 tests pass, and all 19 of this file's tests
are preserved verbatim. The externally used surface (run, serve_inbound,
InboundParams, probe_inbound, relay_rtp, IDLE_POLL_INTERVAL,
OUTBOUND_INVITE_TIMEOUT, OUTBOUND_RING_TIMEOUT) is unchanged, so volte/,
vowifi/ and commands/ needed no edits.

The two OUTBOUND_* constants stay declared in mod.rs rather than moving
into origination.rs: vowifi/mod.rs uses them only from test code, so a
`pub(crate) use` re-export of them reads as unused in a non-test build.
