+++ gsm-sip-bridge/src/ims/agent/origination.rs
@@ -1,0 +1,946 @@
+//! Placing an outbound call on the carrier leg
+//! (specs/025-outbound-calling, specs/029-interruptible-origination-wait).
+//!
+//! Split out of `agent::mod` as the single largest self-contained block there:
+//! an explicit state machine ([`PendingOrigination`]) that the dispatch loop
+//! advances one tick at a time. It replaced a blocking `originate_and_bridge`
+//! that read the carrier socket directly — racing the always-running
+//! client-reader thread (research R2) and wedging the whole loop for up to
+//! ~80s. Carrier responses now arrive through `inbound.rx` like every other
+//! message; everything here is the state they are applied against.
+
+use super::call::{ActiveCall, DialogInfo};
+use super::veth::{spawn_relay, spawn_veth_uas_listener, VethUasResult, VETH_INVITE_TIMEOUT};
+use crate::error::BridgeResult;
+use crate::ims::lifecycle::{BridgedCall, CallStage};
+use crate::ims::sdp::{self, NegotiatedCodec};
+use crate::ims::sip_client::{random_hex, SipResponse};
+use crate::vowifi::control::{reason, write_msg, ControlMessage};
+use chrono::Utc;
+use std::net::{IpAddr, TcpStream, UdpSocket};
+use std::sync::atomic::AtomicBool;
+use std::sync::{mpsc, Arc};
+use std::time::{Duration, Instant};
+
+/// The two carrier-response windows this state machine enforces. They live in
+/// `agent::mod` because `vowifi::mod` cross-checks its own timeouts against
+/// them; see their doc comments there.
+use super::{OUTBOUND_INVITE_TIMEOUT, OUTBOUND_RING_TIMEOUT};
+
+/// How long to wait for a final response to our own CANCEL of an
+/// abandoned INVITE — normally a prompt `487 Request Terminated`, plus the
+/// (legitimate, RFC 3261 §9.1) chance of a `200 OK` racing in from before
+/// the CANCEL arrived at the carrier. Short: a carrier that hasn't reacted
+/// to a CANCEL within a few seconds isn't going to.
+const CANCEL_RESPONSE_TIMEOUT: Duration = Duration::from_secs(5);
+
+/// How long to keep retrying a CANCEL that could not be sent (the transport was
+/// momentarily unavailable) before giving up. Generous: an INVITE we could not
+/// cancel is a phantom-leg liability, so we keep trying well past a transient
+/// blip (greptile PR #35). If the transport is dead this long, the dispatch
+/// loop's own Gm-liveness/renewal machinery is already replacing the session
+/// out from under this attempt anyway.
+const CANCEL_SEND_MAX_WAIT: Duration = Duration::from_secs(30);
+
+/// Hangs up a carrier leg that was answered (200 OK + ACK already
+/// exchanged) but can't be usably bridged for some reason discovered
+/// afterward — sends a BYE to the carrier (reusing `dialog`, built right
+/// after ACK) and `CallEnded` to Agent B over `control`. Found live
+/// (specs/025-outbound-calling review): by the time any of these failures
+/// are discovered, Agent B has very likely *already* answered its own
+/// phone/PBX leg (`Call::make` dispatches the veth INVITE and returns
+/// immediately; `bridge_outbound_leg` answers the phone leg right after,
+/// without waiting for it to connect) — leaving Agent B silent here would
+/// strand a caller who thinks they're connected, on top of leaking the
+/// carrier leg itself. Best-effort on both: there is nothing further to
+/// retry from this point.
+fn hangup_answered_carrier_leg(
+    session: &mut crate::ims::RegisteredSession,
+    control: &mut TcpStream,
+    dialog: &DialogInfo,
+    call_id: &str,
+    reason: &str,
+) {
+    let bye = dialog.build_bye_for(call_id);
+    let _ = session.transport_mut().and_then(|t| t.send(&bye));
+    let _ = write_msg(
+        control,
+        &ControlMessage::CallEnded {
+            call_id: call_id.to_string(),
+            reason: reason.to_string(),
+        },
+    );
+}
+
+/// An outbound origination in flight, held by the dispatch loop across poll
+/// ticks so the wait for the carrier (and then Agent B's veth leg) is
+/// interruptible (specs/029-interruptible-origination-wait).
+pub(super) struct PendingOrigination {
+    step: OriginationStep,
+    /// Correlates carrier responses (by `Call-ID`) and Agent B's `CallEnded`
+    /// (by `call_id`) to *this* attempt — a message naming another call is
+    /// ignored, never acted on (FR-010).
+    call_id: String,
+    from_tag: String,
+    branch: String,
+    invite_cseq: u32,
+    callee_uri: String,
+    route_headers: Vec<String>,
+    via_transport: &'static str,
+    destination: String,
+    /// Connection to Agent B: written to here (`CallRinging`/`CallPlaced`/
+    /// `CallFailed`), read from via `ctrl_rx`'s reader thread.
+    control: TcpStream,
+    /// Spawned in `begin_origination` — *before* the carrier answers, unlike
+    /// the old code that only started reading Agent B once a call was fully
+    /// bridged — so a caller hangup mid-attempt is observable at all.
+    ctrl_rx: mpsc::Receiver<ControlMessage>,
+    rtp_socket: UdpSocket,
+    veth_local_ip: IpAddr,
+    veth_sip_port: u16,
+    wideband: bool,
+    /// While `AwaitingCarrier`: the carrier-response deadline (initially
+    /// `OUTBOUND_INVITE_TIMEOUT`, extended to `OUTBOUND_RING_TIMEOUT` once any
+    /// response arrives). While `AwaitingVeth`: the veth-leg deadline
+    /// (`VETH_INVITE_TIMEOUT`). One field, reset on the transition.
+    deadline: Instant,
+    any_response_seen: bool,
+    ringing_relayed: bool,
+    pub(super) lifecycle: BridgedCall,
+}
+
+/// The waits the old `originate_and_bridge` blocked on, now explicit states.
+/// The shared `Awaiting` prefix is deliberate — each is a distinct thing the
+/// loop is waiting for.
+#[allow(clippy::enum_variant_names)]
+enum OriginationStep {
+    /// INVITE sent; waiting for the carrier's final response.
+    AwaitingCarrier,
+    /// Carrier answered `200 OK` and was ACKed; waiting for Agent B's veth leg.
+    /// The `VETH_INVITE_TIMEOUT` window, previously a blocking
+    /// `veth_rx.recv_timeout`.
+    AwaitingVeth {
+        dialog: DialogInfo,
+        veth_rx: mpsc::Receiver<BridgeResult<VethUasResult>>,
+        /// The codec the carrier answered with, carried to `finish_origination`
+        /// where the transcoding decision (and the "codec we never offered"
+        /// check) is made — kept in the same order as the old blocking path.
+        answer_codec: sdp::NegotiatedCodec,
+    },
+    /// A CANCEL is being sent (abandonment or our own timeout) and we are
+    /// waiting for its outcome — a `487` for the INVITE (expected), or a `200`
+    /// that raced the CANCEL (the carrier answered anyway; we ACK then BYE).
+    /// That outcome arrives on `inbound.rx` like every other response, so it is
+    /// handled at the response arm rather than by a direct socket read that
+    /// would race the client reader (greptile PR #35 / research R2).
+    ///
+    /// `sent` tracks whether the CANCEL actually went out: if the transport was
+    /// momentarily unavailable it is retried each tick, and the response window
+    /// is only armed once it succeeds — otherwise a failed send would drop a
+    /// still-live INVITE, leaking a phantom leg if the carrier later answered
+    /// (greptile PR #35, round 2).
+    AwaitingCancel { sent: bool },
+}
+
+/// Whether a pending origination is still in flight or has resolved this tick.
+pub(super) enum OriginationStatus {
+    /// Keep waiting — the caller retains the `PendingOrigination`.
+    Pending,
+    /// Resolved (failure, timeout, or abandonment). Any teardown and any
+    /// `CallFailed` to Agent B has already been done; the caller drops it.
+    Ended,
+}
+
+/// Best-effort `CallFailed` to Agent B, factored out of the old `fail` closure.
+fn origination_failed(control: &mut TcpStream, call_id: &str, reason: &str) {
+    tracing::warn!(call_id, reason, "outbound: could not place carrier call");
+    let _ = write_msg(
+        control,
+        &ControlMessage::CallFailed {
+            call_id: call_id.to_string(),
+            reason: reason.to_string(),
+        },
+    );
+}
+
+/// Everything `begin_origination` needs about where the veth leg will land.
+pub(super) struct OriginationSetup {
+    pub(super) veth_local_ip: IpAddr,
+    pub(super) veth_sip_port: u16,
+    pub(super) wideband: bool,
+}
+
+/// Builds and sends the carrier INVITE and returns the in-flight state, or
+/// `None` (having told Agent B `CallFailed`) if it could not even be sent. The
+/// front half of the old `originate_and_bridge`, up to and including the
+/// INVITE — plus spawning the Agent B control reader *now* so a mid-attempt
+/// hangup can be heard. The dispatch loop has already sent `CallAttempting`.
+pub(super) fn begin_origination(
+    session: &mut crate::ims::RegisteredSession,
+    mut control: TcpStream,
+    call_id: String,
+    destination: &str,
+    setup: &OriginationSetup,
+) -> Option<PendingOrigination> {
+    // RFC 3608, same simplification `ims::call::run_call` already makes:
+    // subsequent requests in this dialog route via the Service-Route the
+    // registrar returned, computed once here and reused for the INVITE and
+    // (via `DialogInfo::from_uac_response`) the eventual BYE.
+    let route_headers: Vec<String> = session
+        .headers
+        .iter()
+        .filter(|(k, _)| k.eq_ignore_ascii_case("Service-Route"))
+        .map(|(_, v)| format!("Route: {v}"))
+        .collect();
+
+    let rtp_socket = match UdpSocket::bind((session.local_addr.ip(), 0)) {
+        Ok(s) => s,
+        Err(e) => {
+            origination_failed(
+                &mut control,
+                &call_id,
+                &format!("RTP socket bind failed: {e}"),
+            );
+            return None;
+        }
+    };
+    let rtp_port = match rtp_socket.local_addr() {
+        Ok(a) => a.port(),
+        Err(e) => {
+            origination_failed(
+                &mut control,
+                &call_id,
+                &format!("RTP local_addr failed: {e}"),
+            );
+            return None;
+        }
+    };
+
+    let session_id: u64 = rand::random::<u32>() as u64;
+    // Offering, not answering — "prefer wideband when available" for both
+    // carrier paths, rather than reusing `AnswerPreference`'s legacy/cellular
+    // split, which is about the *answer* fallback order (AMR-NB vs. PCMU)
+    // when the far end's own offer lacks AMR-WB — not applicable to what we
+    // ourselves offer here.
+    let offer = sdp::build_offer(
+        session.local_addr.ip(),
+        rtp_port,
+        session_id,
+        sdp::CodecOffer::preferring_wideband(setup.wideband && amr_safe::is_available()),
+    );
+
+    // `;user=phone` (RFC 3261 §19.1.1 / TS 24.229): tells the network this is
+    // a PSTN/mobile number, not a resolvable SIP address — the same header
+    // `ims::call::run_call` adds after finding a bare `sip:` URI reached a
+    // terminating application server that never rang the callee.
+    let callee_uri = format!("{destination}@{};user=phone", session.realm);
+    let from_tag = random_hex(4);
+    let invite_cseq = session.cseq;
+    let via_transport = if session.use_tcp { "TCP" } else { "UDP" };
+    let branch = format!("z9hG4bK{}", random_hex(6));
+
+    let invite = crate::ims::call::build_invite(&crate::ims::call::InviteParts {
+        request_uri: &callee_uri,
+        route_headers: &route_headers,
+        via_transport,
+        local_addr: session.local_addr,
+        contact_addr: session.contact_addr,
+        public_uri: &session.public_uri,
+        callee_uri: &callee_uri,
+        call_id: &call_id,
+        from_tag: &from_tag,
+        cseq: invite_cseq,
+        branch: &branch,
+        body: &offer,
+    });
+
+    tracing::info!(call_id, destination, "outbound: sending INVITE to carrier");
+    let transport = match session.transport_mut() {
+        Ok(t) => t,
+        Err(e) => {
+            origination_failed(
+                &mut control,
+                &call_id,
+                &format!("no carrier transport: {e}"),
+            );
+            return None;
+        }
+    };
+    if let Err(e) = transport.send(&invite) {
+        origination_failed(&mut control, &call_id, &format!("INVITE send failed: {e}"));
+        return None;
+    }
+
+    // Start listening to Agent B now (not after the call is bridged, as the old
+    // code did): this is the whole point of specs/029 — a caller hanging up
+    // while the carrier is still ringing has to reach the dispatch loop, and
+    // this reader is the only path for it.
+    let ctrl_rx = match control.try_clone() {
+        Ok(s) => super::call::spawn_control_reader(s),
+        Err(e) => {
+            origination_failed(
+                &mut control,
+                &call_id,
+                &format!("control connection clone failed: {e}"),
+            );
+            return None;
+        }
+    };
+
+    // Lifecycle from the moment the INVITE is sent: `Offered → Answering`. This
+    // is also what makes the line read as busy to an inbound INVITE arriving
+    // mid-attempt (`Admission::for_current`, FR-011) with no new rule.
+    let mut lifecycle = BridgedCall::new(call_id.clone(), destination.to_string(), None);
+    lifecycle.advance_to(CallStage::Answering);
+
+    Some(PendingOrigination {
+        step: OriginationStep::AwaitingCarrier,
+        call_id,
+        from_tag,
+        branch,
+        invite_cseq,
+        callee_uri,
+        route_headers,
+        via_transport,
+        destination: destination.to_string(),
+        control,
+        ctrl_rx,
+        rtp_socket,
+        veth_local_ip: setup.veth_local_ip,
+        veth_sip_port: setup.veth_sip_port,
+        wideband: setup.wideband,
+        deadline: Instant::now() + OUTBOUND_INVITE_TIMEOUT,
+        any_response_seen: false,
+        ringing_relayed: false,
+        lifecycle,
+    })
+}
+
+impl PendingOrigination {
+    /// Does this carrier response belong to this attempt? Matched by `Call-ID`,
+    /// so it never collides with the Gm keepalive's `OPTIONS` (a different
+    /// `Call-ID`, correlated by `CSeq` at the response arm instead).
+    pub(super) fn matches_response(&self, resp: &SipResponse) -> bool {
+        resp.header("Call-ID")
+            .is_some_and(|id| id.trim() == self.call_id)
+    }
+
+    /// Tell Agent B this attempt failed. Best-effort; Agent B may already have
+    /// moved on (e.g. it initiated an abandonment).
+    fn fail(&mut self, reason: &str) {
+        let call_id = self.call_id.clone();
+        origination_failed(&mut self.control, &call_id, reason);
+    }
+
+    /// Send a CANCEL for a still-pending INVITE and move to `AwaitingCancel`,
+    /// where the response arm handles the `487`/racing-`200` via `inbound.rx`.
+    /// Valid from `AwaitingCarrier`. The origination is kept alive (a short
+    /// `CANCEL_RESPONSE_TIMEOUT` deadline) so a racing answer is still ACKed and
+    /// BYE'd rather than leaking (greptile PR #35).
+    fn begin_cancel(&mut self, session: &mut crate::ims::RegisteredSession) {
+        let sent = self.send_cancel_now(session);
+        self.step = OriginationStep::AwaitingCancel { sent };
+        // Once the CANCEL is out, wait a short window for its response; if it
+        // could not be sent yet, allow a much longer window to keep retrying
+        // rather than dropping a still-live INVITE (greptile PR #35).
+        self.deadline = Instant::now()
+            + if sent {
+                CANCEL_RESPONSE_TIMEOUT
+            } else {
+                CANCEL_SEND_MAX_WAIT
+            };
+    }
+
+    /// Sends CANCEL for a pending outbound INVITE we're giving up on before a
+    /// final response ever arrived (RFC 3261 §9.1) — reusing the original
+    /// INVITE's own branch/CSeq number, since this targets that same
+    /// transaction rather than starting a new one. Found live
+    /// (specs/025-outbound-calling review): abandoning the transaction without
+    /// this left the carrier free to keep ringing the destination for as long
+    /// as *it* was willing to wait, regardless of how long we'd already given
+    /// up.
+    ///
+    /// **Does not read the response** (specs/029, greptile PR #35): the CANCEL's
+    /// outcome — a `487` for the INVITE, or a `200` that raced it — arrives on
+    /// `inbound.rx` and is handled via the `AwaitingCancel` step. The old
+    /// version read the carrier socket directly here, which raced the
+    /// always-running client-reader thread (research R2): if that thread grabbed
+    /// the racing `200`, this timed out and sent no ACK/BYE, leaving a phantom
+    /// carrier leg — the exact failure this whole feature exists to prevent.
+    ///
+    /// Returns whether the CANCEL was actually sent — a `false` means the
+    /// transport was momentarily unavailable and the caller should keep the
+    /// origination alive and retry, rather than assume the INVITE was cancelled.
+    fn send_cancel_now(&mut self, session: &mut crate::ims::RegisteredSession) -> bool {
+        let cancel = crate::ims::call::build_cancel(&crate::ims::call::CancelParts {
+            request_uri: &self.callee_uri,
+            route_headers: &self.route_headers,
+            via_transport: self.via_transport,
+            local_addr: session.local_addr,
+            public_uri: &session.public_uri,
+            callee_uri: &self.callee_uri,
+            call_id: &self.call_id,
+            from_tag: &self.from_tag,
+            cseq: self.invite_cseq,
+            branch: &self.branch,
+        });
+        let Ok(transport) = session.transport_mut() else {
+            return false;
+        };
+        if transport.send(&cancel).is_ok() {
+            tracing::info!(call_id = %self.call_id, "outbound: sent CANCEL for an abandoned INVITE");
+            true
+        } else {
+            false
+        }
+    }
+
+    /// The carrier answered despite our CANCEL — a `200` that raced the `487`.
+    /// ACK it (reusing the INVITE's branch/CSeq, §17.1.1.3) then immediately
+    /// BYE, or the carrier leg would stay up with nothing on our end tracking
+    /// it. Best-effort; there is nothing to retry from here.
+    fn ack_and_bye_racing_answer(
+        &mut self,
+        session: &mut crate::ims::RegisteredSession,
+        resp: &SipResponse,
+    ) {
+        tracing::warn!(
+            call_id = %self.call_id,
+            "outbound: carrier answered despite CANCEL; sending ACK then BYE to hang up"
+        );
+        let to_header = resp.header("To").unwrap_or(&self.callee_uri).to_string();
+        let ack = crate::ims::call::build_ack(&crate::ims::call::AckParts {
+            request_uri: &self.callee_uri,
+            route_headers: &self.route_headers,
+            via_transport: self.via_transport,
+            local_addr: session.local_addr,
+            public_uri: &session.public_uri,
+            to_header: &to_header,
+            call_id: &self.call_id,
+            from_tag: &self.from_tag,
+            cseq: self.invite_cseq,
+            branch: &self.branch,
+        });
+        let _ = session.transport_mut().and_then(|t| t.send(&ack));
+        let bye = crate::ims::call::build_bye(&crate::ims::call::AckParts {
+            request_uri: &self.callee_uri,
+            route_headers: &self.route_headers,
+            via_transport: self.via_transport,
+            local_addr: session.local_addr,
+            public_uri: &session.public_uri,
+            to_header: &to_header,
+            call_id: &self.call_id,
+            from_tag: &self.from_tag,
+            cseq: self.invite_cseq + 1,
+            branch: &format!("z9hG4bK{}", random_hex(6)),
+        });
+        let _ = session.transport_mut().and_then(|t| t.send(&bye));
+    }
+
+    /// Handle a carrier response while awaiting the CANCEL's outcome. Only the
+    /// *INVITE*'s own final resolves the transaction: a `200` on the INVITE
+    /// raced the CANCEL and is ACKed then BYE'd; a `487` (or any other INVITE
+    /// final) means it is dead. The `200` answering the CANCEL request itself
+    /// (CSeq `... CANCEL`) merely confirms the CANCEL — acting on it would end
+    /// tracking too early and miss a racing INVITE answer, leaving a phantom leg
+    /// (greptile PR #35). Provisionals and the CANCEL's own response keep us
+    /// waiting.
+    fn on_cancel_response(
+        &mut self,
+        resp: &SipResponse,
+        session: &mut crate::ims::RegisteredSession,
+    ) -> OriginationStatus {
+        let is_invite = resp
+            .header("CSeq")
+            .and_then(super::ping::cseq_method)
+            .is_some_and(|m| m.eq_ignore_ascii_case("INVITE"));
+        if !is_invite || resp.status < 200 {
+            return OriginationStatus::Pending;
+        }
+        if resp.status == 200 {
+            self.ack_and_bye_racing_answer(session, resp);
+        }
+        OriginationStatus::Ended
+    }
+
+    /// Advance on a carrier response delivered via `inbound.rx`. Returns
+    /// `Ended` (and has already sent `CallFailed`/ACKed as needed) when the
+    /// attempt is resolved as a failure; `Pending` while it is still in flight
+    /// (including the `200 OK → AwaitingVeth` transition).
+    pub(super) fn on_carrier_response(
+        &mut self,
+        resp: &SipResponse,
+        session: &mut crate::ims::RegisteredSession,
+    ) -> OriginationStatus {
+        // Only the carrier-wait phase interprets responses as the INVITE's
+        // outcome. In the other phases a Call-ID-matched response is not a fresh
+        // final to act on (greptile PR #35):
+        match self.step {
+            OriginationStep::AwaitingCancel { .. } => {
+                return self.on_cancel_response(resp, session)
+            }
+            OriginationStep::AwaitingVeth { .. } => {
+                // The INVITE's final was already handled and the veth leg is
+                // being placed. Re-running the answer path would spawn a second
+                // veth listener and send `CallPlaced` twice; ignore the response
+                // instead, exactly as the old blocking code did (it never read
+                // the socket again during the veth wait).
+                //
+                // A retransmitted `2xx` (our ACK was lost, only possible on a
+                // UDP Gm) is therefore not re-ACKed here — and neither is one on
+                // an already-bridged `ActiveCall`, nor on an inbound call
+                // (greptile PR #35 round 4). Retransmitted-2xx re-ACK is a
+                // pre-existing gap orthogonal to this feature and uniform across
+                // both call directions; closing it belongs in a dedicated
+                // change, not piecemeal on only the outbound veth window. Moot
+                // on the TCP Gm transport this deployment uses.
+                tracing::debug!(
+                    call_id = %self.call_id,
+                    status = resp.status,
+                    "ignoring a carrier response after the final was already handled"
+                );
+                return OriginationStatus::Pending;
+            }
+            OriginationStep::AwaitingCarrier => {}
+        }
+
+        // First response of any kind (even a provisional) switches from the
+        // short "did the network acknowledge us at all" window to the long
+        // ring window — the same rule the old blocking origination read
+        // applied internally before this became a poll loop.
+        if !self.any_response_seen {
+            self.any_response_seen = true;
+            self.deadline = Instant::now() + OUTBOUND_RING_TIMEOUT;
+        }
+
+        if resp.status < 200 {
+            tracing::info!(status = resp.status, reason = %resp.reason, "provisional response");
+            // Relay ringback exactly once, and advance the lifecycle to
+            // `PbxRinging` — the transition the old outbound path skipped, so
+            // no successful outbound call ever recorded as reaching `Bridged`
+            // (research R5).
+            if resp.status == 180 && !self.ringing_relayed {
+                self.ringing_relayed = true;
+                self.lifecycle.advance_to(CallStage::PbxRinging);
+                let _ = write_msg(
+                    &mut self.control,
+                    &ControlMessage::CallRinging {
+                        call_id: self.call_id.clone(),
+                    },
+                );
+            }
+            return OriginationStatus::Pending;
+        }
+
+        tracing::info!(call_id = %self.call_id, status = resp.status, reason = %resp.reason, "outbound: final INVITE response");
+
+        if resp.status != 200 {
+            // Non-2xx final: ACK reuses the INVITE's own branch/CSeq
+            // (RFC 3261 §17.1.1.3), best-effort.
+            let ack = crate::ims::call::build_ack(&crate::ims::call::AckParts {
+                request_uri: &self.callee_uri,
+                route_headers: &self.route_headers,
+                via_transport: self.via_transport,
+                local_addr: session.local_addr,
+                public_uri: &session.public_uri,
+                to_header: resp.header("To").unwrap_or(&self.callee_uri),
+                call_id: &self.call_id,
+                from_tag: &self.from_tag,
+                cseq: self.invite_cseq,
+                branch: &self.branch,
+            });
+            let _ = session.transport_mut().and_then(|t| t.send(&ack));
+            self.fail(&format!("{} {}", resp.status, resp.reason));
+            return OriginationStatus::Ended;
+        }
+
+        // 200 OK. Everything from here mirrors the old post-final block, in the
+        // same order (SDP → RTP → ACK → dialog → veth listener → CallPlaced),
+        // so every teardown path keeps its original meaning.
+        let answer = match sdp::parse_answer(&resp.body) {
+            Ok(a) => a,
+            Err(e) => {
+                self.fail(&format!("bad SDP answer: {e}"));
+                return OriginationStatus::Ended;
+            }
+        };
+        if let Err(e) = self.rtp_socket.connect(answer.remote_rtp) {
+            self.fail(&format!("RTP connect failed: {e}"));
+            return OriginationStatus::Ended;
+        }
+
+        let ack_branch = format!("z9hG4bK{}", random_hex(6));
+        let ack = crate::ims::call::build_ack(&crate::ims::call::AckParts {
+            request_uri: &self.callee_uri,
+            route_headers: &self.route_headers,
+            via_transport: self.via_transport,
+            local_addr: session.local_addr,
+            public_uri: &session.public_uri,
+            to_header: resp.header("To").unwrap_or(&self.callee_uri),
+            call_id: &self.call_id,
+            from_tag: &self.from_tag,
+            cseq: self.invite_cseq,
+            branch: &ack_branch,
+        });
+        if let Err(e) = session.transport_mut().and_then(|t| t.send(&ack)) {
+            self.fail(&format!("ACK send failed: {e}"));
+            return OriginationStatus::Ended;
+        }
+        session.cseq = self.invite_cseq + 1;
+
+        let dialog = DialogInfo::from_uac_response(
+            resp,
+            self.route_headers.clone(),
+            &self.callee_uri,
+            &session.public_uri,
+            &self.from_tag,
+            session.cseq,
+            session,
+        );
+
+        // Spawn the veth listener *before* telling Agent B to call in — same
+        // ordering `handle_invite` uses for the inbound direction, so the
+        // listener is guaranteed up by the time Agent B's `Call::make` reaches
+        // it.
+        let veth_rx = match spawn_veth_uas_listener(
+            self.veth_local_ip,
+            self.veth_sip_port,
+            self.wideband,
+        ) {
+            Ok(rx) => rx,
+            Err(e) => {
+                // The carrier leg is already ACKed and up (the `dialog`
+                // proves it) — `fail()` alone only tells Agent B, it does
+                // not hang up the real carrier call.
+                tracing::warn!(call_id = %self.call_id, error = %e, "outbound: veth listener failed");
+                let call_id = self.call_id.clone();
+                hangup_answered_carrier_leg(
+                    session,
+                    &mut self.control,
+                    &dialog,
+                    &call_id,
+                    reason::VETH_LEG_FAILED,
+                );
+                return OriginationStatus::Ended;
+            }
+        };
+
+        if let Err(e) = write_msg(
+            &mut self.control,
+            &ControlMessage::CallPlaced {
+                call_id: self.call_id.clone(),
+            },
+        ) {
+            // Same leak as above: the carrier leg is already up. Best-effort
+            // even though this very write just failed.
+            tracing::warn!(call_id = %self.call_id, error = %e, "outbound: failed to notify Agent B the carrier leg is up");
+            let call_id = self.call_id.clone();
+            hangup_answered_carrier_leg(
+                session,
+                &mut self.control,
+                &dialog,
+                &call_id,
+                reason::TRANSPORT_ERROR,
+            );
+            return OriginationStatus::Ended;
+        }
+
+        self.step = OriginationStep::AwaitingVeth {
+            dialog,
+            veth_rx,
+            answer_codec: answer.codec,
+        };
+        self.deadline = Instant::now() + VETH_INVITE_TIMEOUT;
+        OriginationStatus::Pending
+    }
+}
+
+/// Advance a pending origination on a timer tick (no new carrier response):
+/// watch Agent B's control channel for a caller hangup, enforce the current
+/// deadline, and — once the carrier has answered — pick up Agent B's veth leg.
+/// May take `*pending` (leaving it `None`) when the attempt resolves; returns
+/// the bridged `ActiveCall` only on the success path.
+pub(super) fn tick_pending_origination(
+    pending: &mut Option<PendingOrigination>,
+    session: &mut crate::ims::RegisteredSession,
+) -> Option<ActiveCall> {
+    let mut p = pending.take()?;
+
+    // 1. A caller hangup (or Agent B vanishing) abandons the attempt (FR-003).
+    //    A `CallEnded` naming a different call is ignored, never acted on
+    //    (FR-010).
+    let abandoned = match p.ctrl_rx.try_recv() {
+        Ok(ControlMessage::CallEnded { call_id, .. }) if call_id == p.call_id => {
+            tracing::info!(call_id = %p.call_id, "outbound: caller abandoned the attempt; abandoning the carrier leg");
+            true
+        }
+        Ok(ControlMessage::CallEnded { call_id, .. }) => {
+            tracing::warn!(this = %p.call_id, other = %call_id, "ignoring CallEnded for a different call during origination");
+            false
+        }
+        Ok(other) => {
+            tracing::debug!(message = ?other, "ignoring control message during origination");
+            false
+        }
+        Err(mpsc::TryRecvError::Disconnected) => {
+            tracing::warn!(call_id = %p.call_id, "Agent B control connection dropped during origination; abandoning");
+            true
+        }
+        Err(mpsc::TryRecvError::Empty) => false,
+    };
+    if abandoned {
+        abandon_origination(&mut p, session);
+        // If that left us awaiting the CANCEL's outcome, keep the origination
+        // alive so a racing `200` is still ACKed and BYE'd (greptile PR #35);
+        // otherwise it is fully resolved and dropped.
+        if matches!(p.step, OriginationStep::AwaitingCancel { .. }) {
+            *pending = Some(p);
+        }
+        return None;
+    }
+
+    // 2. Read the veth channel without holding a borrow into `p` across a move.
+    let veth_ready = match &p.step {
+        OriginationStep::AwaitingVeth { veth_rx, .. } => Some(veth_rx.try_recv()),
+        _ => None,
+    };
+    match veth_ready {
+        // Agent B's veth leg arrived — bridge the two legs.
+        Some(Ok(veth_result)) => return finish_origination(p, veth_result, session),
+        // Nothing yet: enforce the veth deadline, else keep waiting.
+        Some(Err(mpsc::TryRecvError::Empty)) => {
+            if Instant::now() >= p.deadline {
+                tracing::warn!(call_id = %p.call_id, "outbound: Agent B's veth call never arrived");
+                hangup_pending_carrier_leg(&mut p, session, reason::VETH_LEG_FAILED);
+            } else {
+                *pending = Some(p);
+            }
+            return None;
+        }
+        Some(Err(mpsc::TryRecvError::Disconnected)) => {
+            tracing::warn!(call_id = %p.call_id, "outbound: veth listener stopped before answering");
+            hangup_pending_carrier_leg(&mut p, session, reason::VETH_LEG_FAILED);
+            return None;
+        }
+        None => {}
+    }
+
+    // 3. Awaiting the carrier or its CANCEL response — enforce the deadline,
+    //    and retry a CANCEL that could not be sent yet.
+    let awaiting_cancel = matches!(p.step, OriginationStep::AwaitingCancel { .. });
+    let cancel_unsent = matches!(p.step, OriginationStep::AwaitingCancel { sent: false });
+
+    if Instant::now() >= p.deadline {
+        if awaiting_cancel {
+            // Best-effort exhausted. Either the CANCEL's response never arrived
+            // within its window, or (worse) the CANCEL could not be sent at all
+            // within the long retry window — in that case the INVITE may linger
+            // at the carrier, but the transport being dead this long means the
+            // session is being re-established anyway.
+            if cancel_unsent {
+                tracing::warn!(call_id = %p.call_id, "outbound: could not send CANCEL within the retry window; giving up");
+            } else {
+                tracing::debug!(call_id = %p.call_id, "outbound: CANCEL response window elapsed");
+            }
+        } else {
+            // AwaitingCarrier timeout — our own give-up. Tell Agent B, then
+            // CANCEL and wait out the outcome via `inbound.rx` (AwaitingCancel),
+            // so a racing answer is cleaned up rather than leaking.
+            tracing::warn!(call_id = %p.call_id, "outbound: no final response from carrier in time");
+            p.fail(&format!(
+                "{}: no final response from carrier",
+                reason::CARRIER_TIMEOUT
+            ));
+            p.begin_cancel(session);
+            *pending = Some(p);
+        }
+        return None;
+    }
+
+    // A CANCEL that failed to send earlier (transport momentarily unavailable)
+    // is retried until it goes out; only then is the response window armed, so
+    // a still-live INVITE is never dropped as if it had been cancelled
+    // (greptile PR #35, round 2).
+    if cancel_unsent && p.send_cancel_now(session) {
+        p.step = OriginationStep::AwaitingCancel { sent: true };
+        p.deadline = Instant::now() + CANCEL_RESPONSE_TIMEOUT;
+    }
+
+    *pending = Some(p);
+    None
+}
+
+/// Abandon an in-flight attempt because the originating caller is gone. Tears
+/// the carrier side down as the current step requires: CANCEL a still-pending
+/// INVITE (→ `AwaitingCancel`, kept alive so a racing answer is cleaned up), or
+/// BYE an already-answered leg. No `CallFailed` is sent — Agent B initiated
+/// this and has already reported `CallerAbandoned`.
+fn abandon_origination(p: &mut PendingOrigination, session: &mut crate::ims::RegisteredSession) {
+    match &p.step {
+        OriginationStep::AwaitingCarrier => p.begin_cancel(session),
+        OriginationStep::AwaitingVeth { .. } => {
+            hangup_pending_carrier_leg(p, session, reason::CALLER_HANGUP);
+        }
+        // Already cancelling — a duplicate `CallEnded`; nothing more to do.
+        OriginationStep::AwaitingCancel { .. } => {}
+    }
+}
+
+/// BYE an already-answered carrier leg for a pending (not-yet-bridged) attempt,
+/// reusing the dialog captured at `200 OK`. Only valid in `AwaitingVeth`.
+fn hangup_pending_carrier_leg(
+    p: &mut PendingOrigination,
+    session: &mut crate::ims::RegisteredSession,
+    reason: &str,
+) {
+    // Disjoint field borrows: `dialog` reads `p.step`, the BYE writes
+    // `p.control`, `call_id` is copied out first — so none of these alias.
+    if let OriginationStep::AwaitingVeth { dialog, .. } = &p.step {
+        let call_id = p.call_id.clone();
+        hangup_answered_carrier_leg(session, &mut p.control, dialog, &call_id, reason);
+    }
+}
+
+/// Bridge a carrier leg (already answered and ACKed) to Agent B's veth leg — the
+/// back half of the old `originate_and_bridge`, run once the veth call arrives.
+/// Consumes `p`. Returns the `ActiveCall`, or `None` after tearing the carrier
+/// leg down if bridging fails.
+fn finish_origination(
+    p: PendingOrigination,
+    veth_result: BridgeResult<VethUasResult>,
+    session: &mut crate::ims::RegisteredSession,
+) -> Option<ActiveCall> {
+    let PendingOrigination {
+        step,
+        call_id,
+        from_tag,
+        destination,
+        control,
+        ctrl_rx,
+        rtp_socket,
+        mut lifecycle,
+        ..
+    } = p;
+    let mut control = control;
+    let OriginationStep::AwaitingVeth {
+        dialog,
+        answer_codec,
+        ..
+    } = step
+    else {
+        // Unreachable: only ever called from the `AwaitingVeth` arm.
+        return None;
+    };
+
+    let veth = match veth_result.map_err(|e| e.to_string()) {
+        Ok(v) => v,
+        Err(e) => {
+            tracing::warn!(call_id, error = %e, "outbound: Agent B's veth call failed");
+            hangup_answered_carrier_leg(
+                session,
+                &mut control,
+                &dialog,
+                &call_id,
+                reason::VETH_LEG_FAILED,
+            );
+            return None;
+        }
+    };
+
+    // `parse_answer` only returns which codec the answer picked
+    // (`NegotiatedCodec`), not a payload type — by RFC 3264, a re-used dynamic
+    // payload type on the answer must mean what *our own offer* said it meant,
+    // so there is nothing to re-parse. Reconstructs the rest from what
+    // `sdp::build_offer` is known to always send.
+    let chosen = match answer_codec {
+        NegotiatedCodec::Pcmu => sdp::ChosenCodec {
+            codec: NegotiatedCodec::Pcmu,
+            payload_type: 0,
+            octet_aligned: false,
+        },
+        NegotiatedCodec::AmrWb => sdp::ChosenCodec {
+            codec: NegotiatedCodec::AmrWb,
+            payload_type: 96,
+            octet_aligned: true,
+        },
+        other => {
+            // Never offered — `sdp::build_offer` only ever lists PCMU/AMR-WB.
+            // Agent B's phone/PBX leg is already answered by this point, so
+            // leaving it stranded on dead air on top of leaking the carrier leg
+            // would compound the failure.
+            tracing::error!(
+                call_id,
+                codec = other.name(),
+                "outbound: carrier answered with a codec we never offered"
+            );
+            hangup_answered_carrier_leg(
+                session,
+                &mut control,
+                &dialog,
+                &call_id,
+                reason::TRANSPORT_ERROR,
+            );
+            return None;
+        }
+    };
+
+    let stop = Arc::new(AtomicBool::new(false));
+    let meter = crate::ims::media_stats::MediaMeter::new();
+    let transcoding = chosen.codec != veth.codec.codec;
+    let relay_result = if transcoding {
+        crate::ims::transcode::spawn_transcoding_relay(
+            rtp_socket,
+            veth.rtp_socket,
+            chosen,
+            veth.codec,
+            stop.clone(),
+            &meter,
+        )
+    } else {
+        spawn_relay(rtp_socket, veth.rtp_socket, stop.clone(), &meter);
+        Ok(())
+    };
+    if let Err(e) = relay_result {
+        tracing::error!(call_id, error = %e, "outbound: failed to start media relay");
+        hangup_answered_carrier_leg(
+            session,
+            &mut control,
+            &dialog,
+            &call_id,
+            reason::TRANSPORT_ERROR,
+        );
+        return None;
+    }
+
+    tracing::info!(
+        call_id,
+        destination,
+        carrier_codec = chosen.codec.name(),
+        transcoding,
+        "outbound: call placed and bridged"
+    );
+
+    // `PbxRinging → Bridged`. Force `PbxRinging` first for the rare carrier
+    // that answers `200` with no `180` in between (`advance_to` refuses the
+    // illegal `Answering → Bridged` jump and would otherwise leave the stage
+    // stuck) — research R5.
+    lifecycle.advance_to(CallStage::PbxRinging);
+    lifecycle.advance_to(CallStage::Bridged);
+
+    Some(ActiveCall {
+        control,
+        ctrl_rx,
+        stop,
+        // Our own from_tag doubles as `to_tag`: `handle_bye`'s
+        // `build_200_ok_bye` only falls back to it when the incoming request's
+        // own To header lacks a tag, which a real in-dialog BYE never does.
+        to_tag: from_tag,
+        dialog,
+        call_id,
+        caller: destination,
+        answered_at: Utc::now(),
+        answered_instant: Instant::now(),
+        meter,
+        lifecycle,
+    })
+}
